// src/util/types.ts
export interface Token {
    address: string;
    decimals: number;
    name: string;
    symbol: string;
    primaryColor: string;
    id?: string;
    c_id?: string; // coingecko id
    icon?: string;
    displayDecimals: number;
    minAmount?: number;
    maxAmount?: number;
    precision?: number;
    website?: string;
    description?: string;
    verisafe_sticker?: string;
    price_usd?: BigNumber | null;
    price_usd_24h_change?: BigNumber | null;
    listed: boolean;
    isStableCoin: boolean;
}
// ...
export interface BlockchainState {
    readonly ethAccount: string;
    readonly wallet: Wallet | null;
    readonly web3State: Web3State;
    readonly tokenBalances: TokenBalance[];
    readonly ethBalance: BigNumber;
    readonly wethTokenBalance: TokenBalance | null;
    readonly gasInfo: GasInfo;
    readonly convertBalanceState: ConvertBalanceState;
    readonly tokenBaseIEO: TokenIEO | null;
    readonly tokenBaseBalanceIEO: TokenBalanceIEO | null;
    readonly tokenBalancesIEO: TokenBalanceIEO[];
}

export interface BZXState {
    readonly iTokensData: iTokenData[];
    readonly TokensList: TokenMetadataBZX[];
    readonly bzxLoadingState: BZXLoadingState;
}

export interface RelayerState {
    readonly orders: UIOrder[];
    readonly userOrders: UIOrder[];
    readonly orderBookState: ServerState;
    readonly marketsStatsState: ServerState;
    readonly marketFillsState: ServerState;
    readonly minOrderExpireTimeOnBooks?: number;
    readonly userIEOOrders?: UIOrder[];
    readonly ieoOrders?: SignedOrder[];
    readonly accountMarketStats?: AccountMarketStat[];
    readonly feeRecipient?: string;
    readonly feePercentage?: number;
    readonly orderExpireTime?: number;
}

export interface SwapState {
    readonly baseToken: Token;
    readonly quoteToken: Token;
    readonly quote?: MarketBuySwapQuote | MarketSellSwapQuote;
    readonly isBuy?: boolean;
    readonly quoteState: SwapQuoteState;
}

export interface UIState {
    readonly notifications: Notification[];
    readonly fills: Fill[];
    readonly marketFills: MarketFill;
    readonly userMarketFills: MarketFill;
    readonly userFills: Fill[];
    readonly hasUnreadNotifications: boolean;
    readonly stepsModal: StepsModalState;
    readonly startTour: boolean;
    readonly orderPriceSelected: BigNumber | null;
    readonly orderBuyPriceSelected: BigNumber | null;
    readonly orderSellPriceSelected: BigNumber | null;
    readonly makerAmountSelected: BigNumber | null;
    readonly sidebarOpen: boolean;
    readonly isAffiliate: boolean;
    readonly openFiatOnRampModal: boolean;
    readonly openFiatOnRampChooseModal: boolean;
    readonly fiatType: 'APPLE_PAY' | 'CREDIT_CARD' | 'DEBIT_CARD' | 'CARDS';
    readonly erc20Theme: Theme;
    readonly erc20Layout: string;
    readonly isDynamicLayout: boolean;
    readonly themeName: string;
    readonly generalConfig?: GeneralConfig;
    readonly userConfigData: UserConfigData | null;
    readonly configData?: ConfigData | null;
    readonly orderSecondsExpirationTime?: BigNumber | null;
}

export interface MarketState {
    readonly currencyPair: CurrencyPair;
    readonly baseToken: Token | null;
    readonly quoteToken: Token | null;
    readonly ethInUsd: BigNumber | null;
    readonly quoteInUsd?: BigNumber | null;
    readonly markets: Market[] | null;
    readonly tokensPrice: TokenPrice[] | null;
    readonly marketStats: RelayerMarketStats | null;
    readonly makerAddresses: string[] | null;
    readonly marketsStats?: RelayerMarketStats[] | null;
    readonly marketMakerStats: MarketMakerStats[];
}

export interface StoreState {
    readonly router: RouterState;
    readonly blockchain: BlockchainState;
    readonly relayer: RelayerState;
    readonly ui: UIState;
    readonly market: MarketState;
    readonly collectibles: CollectiblesState;
    readonly bzx: BZXState;
    readonly swap: SwapState;
    readonly aave: AaveState;
}
// ...
export interface CurrencyPair {
    base: string;
    quote: string;
    config: {
        basePrecision: number;
        pricePrecision: number;
        minAmount: number;
        maxAmount: number;
        quotePrecision: number;
    };
}
export interface CurrencyPairMetaData {
    base: string;
    quote: string;
    config?: {
        basePrecision?: number;
        pricePrecision?: number;
        minAmount?: number;
        maxAmount?: number;
        quotePrecision?: number;
    };
}

export interface Market {
    currencyPair: CurrencyPair;
    price: BigNumber | null;
    spreadInPercentage: BigNumber | null;
    bestAsk: BigNumber | null;
    bestBid: BigNumber | null;
}

export enum NotificationKind {
    CancelOrder = 'CancelOrder',
    Market = 'Market',
    Limit = 'Limit',
    OrderFilled = 'OrderFilled',
    TokenTransferred = 'TokenTransferred',
    LendingComplete = 'LendingComplete',
    UnLendingComplete = 'UnLendingComplete',
    BorrowComplete = 'BorrowComplete',
    RepayComplete = 'RepayComplete',
}
// ...
export interface CollectiblesState {
    readonly allCollectibles: { [tokenId: string]: Collectible };
    readonly isCollectionLoaded: boolean;
    readonly allCollectiblesFetchStatus: AllCollectiblesFetchStatus;
    readonly collectibleSelected: Collectible | null;
    readonly collectionSelected: CollectibleCollection;
}

export interface CollectibleMetadataSource {
    fetchAllUserCollectiblesAsync(userAddress: string, collectibleAddress: string): Promise<Collectible[]>;
    fetchCollectiblesAsync(tokenIds: string[], collectibleAddress: string): Promise<Collectible[]>;
    fetchCollectionAsync(collectibleCollectionAddress: string): Promise<CollectibleCollection | null>;
}

export type ThunkCreator<R = Promise<any>> = ActionCreator<ThunkAction<R, StoreState, ExtraArgument, AnyAction>>;